Moved modVM tests to modAionImpl
