+++ modAionImpl/test/org/aion/zero/impl/vm/AvmTransactionExecutorTest.java
@@ -1,7 +1,9 @@
-package org.aion.vm.avm;
+package org.aion.zero.impl.vm;
 
 import java.util.concurrent.TimeUnit;
-import org.aion.vm.avm.resources.PathManager;
+import org.aion.vm.avm.AvmConfigurations;
+import org.aion.vm.avm.AvmProvider;
+import org.aion.vm.avm.AvmTransactionExecutor;
 import org.aion.vm.avm.schedule.AvmVersionSchedule;
 import org.aion.avm.stub.AvmVersion;
 import org.junit.AfterClass;
@@ -25,8 +27,8 @@ public class AvmTransactionExecutorTest {
     @BeforeClass
     public static void setupClass() {
         AvmVersionSchedule schedule = AvmVersionSchedule.newScheduleForBothVersions(VERSION_1_FORK, VERSION_2_FORK, TOLERANCE);
-        AvmConfigurations.initializeConfigurationsAsReadOnly(schedule, PathManager.fetchProjectRootDir());
-        projectRootDir = PathManager.fetchProjectRootDir();
+        AvmConfigurations.initializeConfigurationsAsReadAndWriteable(schedule, AvmPathManager.getPathOfProjectRootDirectory());
+        projectRootDir = AvmPathManager.getPathOfProjectRootDirectory();
     }
 
     @AfterClass
